requirements.txt
paho-mqtt==1.6.1
python-dotenv==1.0.0
tqdm==4.66.1
python-dotenv==1.0.0